+++ contracts/DiamondCut.sol
@@ -7,7 +7,6 @@ import {IDiamondCut} from "./interfaces/IDiamondCut.sol"; // Interface for Diamo
 /// @notice This contract is responsible for adding, removing, or replacing facets in a Diamond contract.
 /// @dev The DiamondCut contract implements the `IDiamondCut` interface, enabling modification of facets dynamically.
 contract DiamondCut is IDiamondCut {
-    
     // map facet addresses to function selectors
     mapping(address => bytes4[]) public facetFunctionSelectors;
 
@@ -18,7 +17,7 @@ contract DiamondCut is IDiamondCut {
     event FacetAdded(address indexed facetAddress, bytes4[] functionSelectors);
     event FacetReplaced(address indexed oldFacetAddress, address indexed newFacetAddress, bytes4[] functionSelectors); // Declare the event
     event FacetRemoved(address indexed facetAddress, bytes4[] functionSelectors);
-    event FacetInitialized(address indexed facetAddress);  // Event for facet initialization
+    event FacetInitialized(address indexed facetAddress); // Event for facet initialization
 
     /// @notice Adds, replaces, or removes facets in the Diamond contract
     /// @dev This function allows for dynamic updates to the Diamond contract's facets.
@@ -27,13 +26,9 @@ contract DiamondCut is IDiamondCut {
     /// @param _diamondCut An array of `FacetCut` structs that define the facets and the action to perform on each
     /// @param _init The address of the contract to initialize after facet updates (optional)
     /// @param _calldata The calldata to be passed to the initializer contract (if `_init` is not `address(0)`)
-    function diamondCut(
-        FacetCut[] calldata _diamondCut,
-        address _init,
-        bytes calldata _calldata
-    ) external override {
+    function diamondCut(FacetCut[] calldata _diamondCut, address _init, bytes calldata _calldata) external override {
         // Loop through each facet cut and execute the corresponding action
-        for (uint i = 0; i < _diamondCut.length; i++) {
+        for (uint256 i = 0; i < _diamondCut.length; i++) {
             FacetCut memory cut = _diamondCut[i];
             if (cut.action == FacetCutAction.Add) {
                 _addFacet(cut.facetAddress, cut.functionSelectors);
@@ -46,9 +41,9 @@ contract DiamondCut is IDiamondCut {
 
         // If an initializer contract is provided, perform the delegatecall for initialization
         if (_init != address(0)) {
-            (bool success, ) = _init.delegatecall(_calldata);
+            (bool success,) = _init.delegatecall(_calldata);
             require(success, "DiamondCut: Initialization failed");
-            emit FacetInitialized(_init);  // Emit the event for facet initialization
+            emit FacetInitialized(_init); // Emit the event for facet initialization
         }
     }
 
@@ -60,7 +55,7 @@ contract DiamondCut is IDiamondCut {
     function _addFacet(address _facetAddress, bytes4[] memory _functionSelectors) internal {
         require(_facetAddress != address(0), "DiamondCut: Facet address cannot be zero");
 
-        for (uint i = 0; i < _functionSelectors.length; i++) {
+        for (uint256 i = 0; i < _functionSelectors.length; i++) {
             bytes4 selector = _functionSelectors[i];
             // Ensure the selector isn't already in use by another facet
             require(selectorToFacet[selector] == address(0), "DiamondCut: Function selector already in use");
@@ -83,7 +78,7 @@ contract DiamondCut is IDiamondCut {
     function _replaceFacet(address _facetAddress, bytes4[] memory _functionSelectors) internal {
         require(_facetAddress != address(0), "DiamondCut: Facet address cannot be zero");
 
-        for (uint i = 0; i < _functionSelectors.length; i++) {
+        for (uint256 i = 0; i < _functionSelectors.length; i++) {
             bytes4 selector = _functionSelectors[i];
             address existingFacet = selectorToFacet[selector];
             require(existingFacet != address(0), "DiamondCut: Selector not found");
@@ -99,7 +94,7 @@ contract DiamondCut is IDiamondCut {
             facetFunctionSelectors[_facetAddress].push(selector);
         }
 
-        // emit FacetReplaced(existingFacet, _facetAddress, _functionSelectors); 
+        // emit FacetReplaced(existingFacet, _facetAddress, _functionSelectors);
     }
 
     /// @notice Internal function to remove a facet from the Diamond contract
@@ -110,7 +105,7 @@ contract DiamondCut is IDiamondCut {
     function _removeFacet(address _facetAddress, bytes4[] memory _functionSelectors) internal {
         require(_facetAddress != address(0), "DiamondCut: Facet address cannot be zero");
 
-        for (uint i = 0; i < _functionSelectors.length; i++) {
+        for (uint256 i = 0; i < _functionSelectors.length; i++) {
             bytes4 selector = _functionSelectors[i];
             require(selectorToFacet[selector] == _facetAddress, "DiamondCut: Function selector not found");
 
@@ -130,7 +125,7 @@ contract DiamondCut is IDiamondCut {
     /// @param _selector The function selector to remove
     function _removeFromFacetSelectorList(address _facetAddress, bytes4 _selector) internal {
         bytes4[] storage selectors = facetFunctionSelectors[_facetAddress];
-        for (uint i = 0; i < selectors.length; i++) {
+        for (uint256 i = 0; i < selectors.length; i++) {
             if (selectors[i] == _selector) {
                 selectors[i] = selectors[selectors.length - 1];
                 selectors.pop();
